% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat_server}
\alias{chat_server}
\title{A chat module for Shiny apps - server}
\usage{
chat_server(
  id,
  chat_user,
  db_connection = NULL,
  db_table_name = "chat_data",
  rds_path = NULL,
  csv_path = NULL,
  invalidateDSMillis = 1000,
  pretty = TRUE,
  nlast = 100
)
}
\arguments{
\item{id}{The id of the module.}

\item{chat_user}{The user name that should be displayed next to the message.}

\item{db_connection}{A database connection object, created using the \code{DBI} package. If provided, the chat messages will be stored in a database table.}

\item{db_table_name}{he name of the database table to use for storing the chat messages. If \code{db_connection} is provided, this parameter is required.}

\item{rds_path}{The path to an RDS file to use for storing the chat messages. If provided, the chat messages will be stored in an RDS file.}

\item{csv_path}{The path to an csv file to use for storing the chat messages. If provided, the chat messages will be stored in an csv file.}

\item{invalidateDSMillis}{The milliseconds to wait before the data source is read again. The default is 1 second.}

\item{pretty}{Logical that determines if the date should be displayed in a pretty format}

\item{nlast}{The number of last messages to be read in and displayed}
}
\value{
the reactive values \code{chat_rv} with all the chat information
}
\description{
Creates the server logic for the chat module, which handles adding new messages to the database or RDS file, and retrieving messages to display
}
