% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat_ui}
\alias{chat_ui}
\title{A chat module for Shiny apps - UI}
\usage{
chat_ui(id, height = "300px", width = "500px")
}
\arguments{
\item{id}{The id of the module}

\item{height}{The height of the chat display area. Default is 300px.}

\item{width}{The width of the chat display area. Default is 500px.}
}
\description{
Creates the user interface for the chat module, which includes a chat message display area, a text input field for entering new messages, and a send button.
}
