% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSVConnection.R
\name{CSVConnection}
\alias{CSVConnection}
\title{CSVConnection R6 Class}
\value{
The full dataset

Save a message to data source
}
\description{
CSVConnection R6 Class

CSVConnection R6 Class
}
\details{
An R6 class representing a connection to a CSV file for the chat module.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{csv_path}}{The path to the CSV file.}

\item{\code{nlast}}{The number of messages to be read in and displayed.

Initialize the R6 Object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CSVConnection-new}{\code{CSVConnection$new()}}
\item \href{#method-CSVConnection-get_data}{\code{CSVConnection$get_data()}}
\item \href{#method-CSVConnection-insert_message}{\code{CSVConnection$insert_message()}}
\item \href{#method-CSVConnection-clone}{\code{CSVConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CSVConnection-new"></a>}}
\if{latex}{\out{\hypertarget{method-CSVConnection-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVConnection$new(csv_path, nlast = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{csv_path}}{The path to the csv file.}

\item{\code{nlast}}{The number of messages to be read-in.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CSVConnection-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-CSVConnection-get_data}{}}}
\subsection{Method \code{get_data()}}{
Reads the full dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVConnection$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CSVConnection-insert_message"></a>}}
\if{latex}{\out{\hypertarget{method-CSVConnection-insert_message}{}}}
\subsection{Method \code{insert_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVConnection$insert_message(message, user, time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{The message to be stores}

\item{\code{user}}{The user who entered the message}

\item{\code{time}}{The time when message was submitted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CSVConnection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CSVConnection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CSVConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
