% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBConnection.R
\name{DBConnection}
\alias{DBConnection}
\title{DBConnection R6 Class}
\value{
The full dataset

Save a message to data source
}
\description{
DBConnection R6 Class

DBConnection R6 Class
}
\details{
An R6 class representing a connection to a database for the chat module.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connection}}{A database connection object, created using a package such as RSQLite.}

\item{\code{table}}{The table that contains the chat information.


Initialize the R6 Object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DBConnection-new}{\code{DBConnection$new()}}
\item \href{#method-DBConnection-get_data}{\code{DBConnection$get_data()}}
\item \href{#method-DBConnection-insert_message}{\code{DBConnection$insert_message()}}
\item \href{#method-DBConnection-clone}{\code{DBConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBConnection-new"></a>}}
\if{latex}{\out{\hypertarget{method-DBConnection-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBConnection$new(connection, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{DB connection}

\item{\code{table}}{Table name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBConnection-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-DBConnection-get_data}{}}}
\subsection{Method \code{get_data()}}{
Reads the full dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBConnection$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBConnection-insert_message"></a>}}
\if{latex}{\out{\hypertarget{method-DBConnection-insert_message}{}}}
\subsection{Method \code{insert_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBConnection$insert_message(message, user, time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{The message to be stores}

\item{\code{user}}{The user who entered the message}

\item{\code{time}}{The time when message was submitted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBConnection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DBConnection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
