% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDSConnection.R
\name{RDSConnection}
\alias{RDSConnection}
\title{RDSConnection R6 Class}
\value{
The full dataset

Save a message to data source
}
\description{
RDSConnection R6 Class

RDSConnection R6 Class
}
\details{
An R6 class representing a connection to a rds file for the chat module.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rds_path}}{The path to the rds file.

Initialize the R6 Object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RDSConnection-new}{\code{RDSConnection$new()}}
\item \href{#method-RDSConnection-get_data}{\code{RDSConnection$get_data()}}
\item \href{#method-RDSConnection-insert_message}{\code{RDSConnection$insert_message()}}
\item \href{#method-RDSConnection-clone}{\code{RDSConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RDSConnection-new"></a>}}
\if{latex}{\out{\hypertarget{method-RDSConnection-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDSConnection$new(rds_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rds_path}}{The path to the rds file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RDSConnection-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-RDSConnection-get_data}{}}}
\subsection{Method \code{get_data()}}{
Reads the full dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDSConnection$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RDSConnection-insert_message"></a>}}
\if{latex}{\out{\hypertarget{method-RDSConnection-insert_message}{}}}
\subsection{Method \code{insert_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDSConnection$insert_message(message, user, time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{The message to be stores}

\item{\code{user}}{The user who entered the message}

\item{\code{time}}{The time when message was submitted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RDSConnection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RDSConnection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RDSConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
