% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render_msg_divs}
\alias{render_msg_divs}
\title{Render the messages for the chat}
\usage{
render_msg_divs(texts, users)
}
\arguments{
\item{texts}{a character vector with the texts}

\item{users}{a character vector with the users}
}
\value{
The HTML code containing the chat messages
}
\description{
Render the messages for the chat
}
\examples{

render_msg_divs(c("hello", "hello"), c("user1", "user2"))


}
