% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{updateChatTextInput}
\alias{updateChatTextInput}
\title{A function to update the chat textInput}
\usage{
updateChatTextInput(session = getDefaultReactiveDomain(), id, value)
}
\arguments{
\item{session}{The shiny session.}

\item{id}{The id of the module.}

\item{value}{The new value that should be shown in the chat textInput.}
}
\description{
Updates the value of the chat textInput
}
